+++ src/runtime/exceptions.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Reflection;
 using System.Runtime.InteropServices;
+using System.Text;
 
 namespace Python.Runtime
 {
@@ -71,14 +72,16 @@ namespace Python.Runtime
                 return Exceptions.RaiseTypeError("invalid object");
             }
 
-            string message = string.Empty;
-            if (e.Message != string.Empty)
+            string message = e.ToString();
+            string fullTypeName = e.GetType().FullName;
+            string prefix = fullTypeName + ": ";
+            if (message.StartsWith(prefix))
             {
-                message = e.Message;
+                message = message.Substring(prefix.Length);
             }
-            if (!string.IsNullOrEmpty(e.StackTrace))
+            else if (message.StartsWith(fullTypeName))
             {
-                message = message + "\n" + e.StackTrace;
+                message = message.Substring(fullTypeName.Length);
             }
             return Runtime.PyUnicode_FromString(message);
         }
@@ -181,6 +184,7 @@ namespace Python.Runtime
 
             if (e.InnerException != null)
             {
+                // Note: For an AggregateException, InnerException is only the first of the InnerExceptions.
                 IntPtr cause = CLRObject.GetInstHandle(e.InnerException);
                 Marshal.WriteIntPtr(ob, ExceptionOffset.cause, cause);
             }
@@ -291,6 +295,20 @@ namespace Python.Runtime
         }
 
         /// <summary>
+        /// When called after SetError, sets the cause of the error.
+        /// </summary>
+        /// <param name="cause">The cause of the current error</param>
+        public static void SetCause(PythonException cause)
+        {
+            var currentException = new PythonException();
+            currentException.Normalize();
+            cause.Normalize();
+            Runtime.XIncref(cause.PyValue);
+            Runtime.PyException_SetCause(currentException.PyValue, cause.PyValue);
+            currentException.Restore();
+        }
+
+        /// <summary>
         /// ErrorOccurred Method
         /// </summary>
         /// <remarks>
@@ -368,17 +386,31 @@ namespace Python.Runtime
         // Internal helper methods for common error handling scenarios.
         //====================================================================
 
+        /// <summary>
+        /// Raises a TypeError exception and attaches any existing exception as its cause.
+        /// </summary>
+        /// <param name="message">The exception message</param>
+        /// <returns><c>IntPtr.Zero</c></returns>
         internal static IntPtr RaiseTypeError(string message)
         {
+            PythonException previousException = null;
+            if (ErrorOccurred())
+            {
+                previousException = new PythonException();
+            }
             Exceptions.SetError(Exceptions.TypeError, message);
+            if (previousException != null)
+            {
+                SetCause(previousException);
+            }
             return IntPtr.Zero;
         }
 
         // 2010-11-16: Arranged in python (2.6 & 2.7) source header file order
         /* Predefined exceptions are
-           puplic static variables on the Exceptions class filled in from
+           public static variables on the Exceptions class filled in from
            the python class using reflection in Initialize() looked up by
-		   name, not posistion. */
+		   name, not position. */
         public static IntPtr BaseException;
         public static IntPtr Exception;
         public static IntPtr StopIteration;